elitecart-backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
mysql-connector-python==8.1.0
pydantic
python-dotenv==1.0.0
werkzeug==2.3.7
